Refactor: Comprehensive SOLID Refactoring, Cleanup, and Modernization
This commit applies extensive refactoring to the UnoraLaunchpad codebase, focusing on improving structure, maintainability, and adherence to SOLID principles.

Key Changes:

1.  **Decoupled `MainWindow.xaml.cs`**:
    *   I extracted core responsibilities from `MainWindow.xaml.cs` into a new service layer:
        *   `Services/SettingsService.cs` (manages settings loading, saving, theme, window dimensions)
        *   `Services/UpdateService.cs` (handles launcher and game file updates, progress reporting)
        *   `Services/LaunchService.cs` (manages game client launching, memory patching, DLL injection)
        *   `Services/NavigationService.cs` (handles opening of dialogs/windows like PatchNotes, Settings)
        *   `Services/UserNotifierService.cs` (standardizes MessageBox interactions)
    *   I created utility classes in `Utils/`:
        *   `LoggingService.cs` (centralized file-based logging)
        *   `FileHashHelper.cs` (MD5 calculation)
        *   `FormattingHelper.cs` (byte/speed string formatting)

2.  **Applied SOLID Principles (DIP & ISP)**:
    *   I introduced interfaces for all service classes in `Interfaces/` (e.g., `ISettingsService`, `IUpdateService`).
    *   I updated `MainWindow.xaml.cs` and inter-service dependencies to rely on these abstractions rather than concrete implementations, promoting looser coupling.

3.  **Standardized Error Handling & Logging**:
    *   I consistently integrated `LoggingService.cs` for detailed exception and event logging throughout the application.
    *   User-facing error messages are now presented via `IUserNotifierService` and are more generic, guiding you to check logs for technical details.

4.  **Improved Resource Management**:
    *   I reviewed and ensured correct implementation of `IDisposable` in `RuntimePatcher.cs` and `SuspendedProcess.cs`.
    *   I corrected static resource initialization (`HttpClient`, Polly policy) in `UnoraClient.cs` to use a static constructor.
    *   I ensured `NotifyIcon` in `MainWindow.xaml.cs` is disposed of reliably.
    *   I confirmed correct use of `using` statements for disposable objects in critical paths.

5.  **Applied DRY (Don't Repeat Yourself) Principle**:
    *   I identified and refactored duplicated code blocks in `SettingsService`, `LaunchService`, `UserNotifierService`, `MainWindow.xaml.cs`, and `RuntimePatcher.cs` by extracting common logic into private helper methods.

6.  **Folder & Namespace Organization (Partial)**:
    *   I created `Models/` folder and moved data classes (`Settings.cs`, `FileDetail.cs`, `GameApiRoutes.cs`, `GameUpdate.cs`) into it, updating namespaces and references.
    *   I moved `UnoraClient.cs` into the `Services/` folder and updated its namespace.
    *   The planned move of `RelayCommand.cs` to `Utils/` was not completed due to turn limits.
    *   A full reorganization of UI elements (Views/Windows) was deferred due to complexity and risks.

7.  **Code Clarity & Documentation (Partial)**:
    *   I added XML documentation to most new service/utility classes and some core classes.
    *   Persistent issues prevented comprehensive documentation for several files (`Settings.cs`, `FileService.cs`, `ClientVersion.cs`, `GameUpdate.cs`, `RelayCommand.cs`, and all UI classes). This remains an area for future improvement.

Known Issues & Limitations:

*   **Incomplete Documentation**: As mentioned, several files lack XML documentation due to problems I encountered.
*   **`FileService.cs`**: This file proved difficult to modify. It remains largely in its original state, and some parts of the app still use it statically. Ideally, it would implement an `IFileService` and its static usage would be refactored.
*   **`RelayCommand.cs` Location**: Still in the project root.
*   **Testing**: All changes are based on static analysis and conceptual review. No automated or manual runtime tests were performed.

Overall, this refactoring effort has significantly modernized the codebase, making it more modular, testable (with the introduction of interfaces), and maintainable.
