+++ UnoraLaunchpad/App.xaml.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Windows;
+using UnoraLaunchpad.Models; // Required for Settings
 
 namespace UnoraLaunchpad
 {
