+++ UnoraLaunchpad/GameUpdateDetailView.xaml.cs
@@ -5,6 +5,7 @@ using System.Windows;
 using System.Windows.Input;
 using System.Windows.Media.Animation;
 using Microsoft.Xaml.Behaviors.Core;
+using UnoraLaunchpad.Models; // Required for GameUpdate, Settings
 
 namespace UnoraLaunchpad;
