+++ UnoraLaunchpad/Services/LaunchService.cs
@@ -6,9 +6,11 @@ using System.Net;
 using System.Net.Sockets;
 using System.Threading.Tasks;
 using System.Windows; // For MessageBox if UserNotifierService is not used directly for all messages
-using UnoraLaunchpad.Definitions; // For Settings, SuspendedProcess, etc.
+using UnoraLaunchpad.Models;    // For Settings
+using UnoraLaunchpad.Launcher;  // For ClientVersion, SuspendedProcess, ProcessMemoryStream, RuntimePatcher
 using UnoraLaunchpad.Interfaces; // Required for ILaunchService
-using UnoraLaunchpad.Utils;     // For NativeMethods if still used directly by any moved logic, LoggingService
+using UnoraLaunchpad.Utils;     // For NativeMethods, LoggingService
+// UnoraLaunchpad.Definitions is not directly used now, CONSTANTS etc. are in their respective places.
 
 namespace UnoraLaunchpad.Services
 {
@@ -105,8 +107,8 @@ namespace UnoraLaunchpad.Services
         private void PatchClient(SuspendedProcess process, IPAddress serverIPAddress, int serverPort)
         {
             // Use CurrentSettings property
-            using var stream = new ProcessMemoryStream(process.ProcessId);
-            using var patcher = new RuntimePatcher(ClientVersion.Version741, stream, true); // Assumes ClientVersion.Version741 is accessible
+            using var stream = new ProcessMemoryStream(process.ProcessId); // From UnoraLaunchpad.Launcher
+            using var patcher = new RuntimePatcher(ClientVersion.Version741, stream, true); // From UnoraLaunchpad.Launcher
 
             if (serverIPAddress == null)
             {
