+++ UnoraLaunchpad/SettingsWindow.xaml.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
+using UnoraLaunchpad.Models; // Required for Settings
 
 namespace UnoraLaunchpad;
