+++ UnoraLaunchpad/UpdateLockWIndow.xaml.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Windows;
+using UnoraLaunchpad.Models; // Required for Settings
 
 namespace UnoraLaunchpad
 {
